Remove AbstractScheduler
